backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-WTF==1.2.2
WTForms==3.2.1
Flask-JWT-Extended==4.5.3
bcrypt==4.3.0
redis==5.0.1
PyMySQL==1.1.0
DBUtils==3.0.2
python-dotenv==1.0.0
python-json-logger==2.0.7
firebase-admin==6.2.0
razorpay==1.4.1
setuptools<81  # Fix deprecation warning
celery==5.5.3
click==8.2.1
Flask-Limiter==3.13
argon2-cffi==25.1.0